Add Handlebars templates, overhaul game start
This overhaul adds the ability to render pages as Handlebars templates.
This does not use built-in Rocket template handling and instead has the
templates baked into the binary.

This change also improves the way gamestate is accessed: the guards
are now contained in a struct that implements FromRequest, so that you
can add a Game or MutGame argument to the request handler to automatically
get a reference to the current game, and fail the request if there is no
current game.

The above two capabilities are used to create a new Start Game page, as
well as corresponding backend functions for setting a time to derby and
team rosters.